refactor(qwen3-ssm): split the TransformerLayer impl into trait_layer.rs
mod.rs hit 507 LoC (over the 500 cap) after the wyN kernel-handle field. Moves the
`impl TransformerLayer for Qwen3SsmLayer` block to a sibling `trait_layer.rs`,
matching the existing trait_decode/trait_prefill idiom in this module. mod.rs 507 -> 236.
No behaviour change; cargo fmt + clippy clean.
