+++ crates/spark-model/src/layers/qwen3_ssm/mod.rs
@@ -18,9 +18,7 @@ use anyhow::Result;
 use spark_runtime::gpu::{DevicePtr, GpuBackend, KernelHandle};
 use spark_runtime::kv_cache::PagedKvCache;
 
-use crate::layer::{
-    ForwardContext, GdnPrefillBuffers, LayerState, SsmLayerState, TransformerLayer,
-};
+use crate::layer::{ForwardContext, GdnPrefillBuffers, LayerState, SsmLayerState};
 use crate::layers::FfnComponent;
 use crate::layers::ops;
 use crate::weight_map::{DenseWeight, Fp8Weight, QuantizedWeight, SsmWeights};
@@ -221,6 +219,7 @@ mod trait_decode_batched;
 mod trait_decode_batched_conv_gdn;
 mod trait_decode_batched_conv_gdn_wyn;
 mod trait_decode_multi_seq;
+mod trait_layer;
 mod trait_prefill;
 mod trait_prefill_gdn;
 mod trait_prefill_helper;
@@ -230,278 +229,6 @@ mod trait_prefill_proj;
 mod trait_prefill_recur;
 
 // ── TransformerLayer impl (delegates to per-file inherent _inner methods) ──
-impl TransformerLayer for Qwen3SsmLayer {
-    fn decode(
-        &self,
-        hidden: DevicePtr,
-        residual: DevicePtr,
-        state: &mut dyn LayerState,
-        kv_cache: &mut PagedKvCache,
-        seq_len: usize,
-        block_table: &mut Vec<u32>,
-        disk_block_ids: &mut Vec<u32>,
-        disk_last_offloaded_per_layer: &mut Vec<u32>,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.decode_inner(
-            hidden,
-            residual,
-            state,
-            kv_cache,
-            seq_len,
-            block_table,
-            disk_block_ids,
-            disk_last_offloaded_per_layer,
-            ctx,
-            stream,
-        )
-    }
-
-    fn decode_batched(
-        &self,
-        hidden: DevicePtr,
-        residual: DevicePtr,
-        num_tokens: usize,
-        state: &mut dyn LayerState,
-        kv_cache: &mut PagedKvCache,
-        seq_len: usize,
-        block_table: &mut Vec<u32>,
-        disk_block_ids: &mut Vec<u32>,
-        disk_last_offloaded_per_layer: &mut Vec<u32>,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.decode_batched_inner(
-            hidden,
-            residual,
-            num_tokens,
-            state,
-            kv_cache,
-            seq_len,
-            block_table,
-            disk_block_ids,
-            disk_last_offloaded_per_layer,
-            ctx,
-            stream,
-        )
-    }
-
-    fn decode_multi_seq<'a, 'b: 'a>(
-        &self,
-        hidden: DevicePtr,
-        residual: DevicePtr,
-        num_seqs: usize,
-        states: &'a mut [&'b mut (dyn LayerState + 'static)],
-        kv_cache: &mut PagedKvCache,
-        seq_lens: &[usize],
-        block_tables: &[Vec<u32>],
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.decode_multi_seq_inner(
-            hidden,
-            residual,
-            num_seqs,
-            states,
-            kv_cache,
-            seq_lens,
-            block_tables,
-            ctx,
-            stream,
-        )
-    }
-
-    fn prefill(
-        &self,
-        hidden: DevicePtr,
-        residual: DevicePtr,
-        num_tokens: usize,
-        state: &mut dyn LayerState,
-        kv_cache: &mut PagedKvCache,
-        seq_len_start: usize,
-        block_table: &mut Vec<u32>,
-        disk_block_ids: &mut Vec<u32>,
-        disk_last_offloaded_per_layer: &mut Vec<u32>,
-        kv_write_start: usize,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.prefill_inner(
-            hidden,
-            residual,
-            num_tokens,
-            state,
-            kv_cache,
-            seq_len_start,
-            block_table,
-            disk_block_ids,
-            disk_last_offloaded_per_layer,
-            kv_write_start,
-            ctx,
-            stream,
-        )
-    }
-
-    fn is_ssm_layer(&self) -> bool {
-        self.is_ssm_layer_inner()
-    }
-
-    fn prefill_phase1(
-        &self,
-        hidden: DevicePtr,
-        residual: DevicePtr,
-        num_tokens: usize,
-        state: &mut dyn LayerState,
-        kv_cache: &mut PagedKvCache,
-        seq_len_start: usize,
-        block_table: &mut Vec<u32>,
-        disk_block_ids: &mut Vec<u32>,
-        disk_last_offloaded_per_layer: &mut Vec<u32>,
-        kv_write_start: usize,
-        gdn_bufs: &GdnPrefillBuffers,
-        token_offset: usize,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.prefill_phase1_inner(
-            hidden,
-            residual,
-            num_tokens,
-            state,
-            kv_cache,
-            seq_len_start,
-            block_table,
-            disk_block_ids,
-            disk_last_offloaded_per_layer,
-            kv_write_start,
-            gdn_bufs,
-            token_offset,
-            ctx,
-            stream,
-        )
-    }
-
-    fn prefill_phase1_proj_batched(
-        &self,
-        hidden_stacked: DevicePtr,
-        residual_stacked: DevicePtr,
-        total_tokens: usize,
-        gdn_bufs: &GdnPrefillBuffers,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.prefill_phase1_proj_batched_inner(
-            hidden_stacked,
-            residual_stacked,
-            total_tokens,
-            gdn_bufs,
-            ctx,
-            stream,
-        )
-    }
-
-    fn prefill_phase1_conv1d_one(
-        &self,
-        state: &mut dyn LayerState,
-        token_offset: usize,
-        len: usize,
-        gdn_bufs: &GdnPrefillBuffers,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.prefill_phase1_conv1d_one_inner(state, token_offset, len, gdn_bufs, ctx, stream)
-    }
-
-    fn prefill_phase1_l2_batched(
-        &self,
-        total_tokens: usize,
-        gdn_bufs: &GdnPrefillBuffers,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.prefill_phase1_l2_batched_inner(total_tokens, gdn_bufs, ctx, stream)
-    }
-
-    fn prefill_gdn_full(
-        &self,
-        state: &mut dyn LayerState,
-        gdn_bufs: &GdnPrefillBuffers,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.prefill_gdn_full_inner(state, gdn_bufs, ctx, stream)
-    }
-
-    fn prefill_gdn_full_batched(
-        &self,
-        h_state_ptrs: DevicePtr,
-        gdn_bufs: &GdnPrefillBuffers,
-        batch_size: u32,
-        chunk_len: u32,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.prefill_gdn_full_batched_inner(
-            h_state_ptrs,
-            gdn_bufs,
-            batch_size,
-            chunk_len,
-            ctx,
-            stream,
-        )
-    }
-
-    fn prefill_gdn_full_batched_fla_varlen(
-        &self,
-        h_state_ptrs: DevicePtr,
-        gdn_bufs: &GdnPrefillBuffers,
-        batch_size: u32,
-        cu_seqlens: DevicePtr,
-        max_num_chunks: u32,
-        total_nt: usize,
-        max_seqlen: u32,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<bool> {
-        self.prefill_gdn_full_batched_fla_varlen_inner(
-            h_state_ptrs,
-            gdn_bufs,
-            batch_size,
-            cu_seqlens,
-            max_num_chunks,
-            total_nt,
-            max_seqlen,
-            ctx,
-            stream,
-        )
-    }
-
-    fn prefill_phase3(
-        &self,
-        hidden: DevicePtr,
-        residual: DevicePtr,
-        num_tokens: usize,
-        gdn_bufs: &GdnPrefillBuffers,
-        token_offset: usize,
-        ctx: &ForwardContext,
-        stream: u64,
-    ) -> Result<()> {
-        self.prefill_phase3_inner(
-            hidden,
-            residual,
-            num_tokens,
-            gdn_bufs,
-            token_offset,
-            ctx,
-            stream,
-        )
-    }
-
-    fn alloc_state(&self, gpu: &dyn GpuBackend) -> Result<Box<dyn LayerState>> {
-        self.alloc_state_inner(gpu)
-    }
-}
 
 #[cfg(test)]
 mod tests;
